// __generated__/__types__.tsx
export type Maybe<T> = T;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
};

export type Analytics = {
  __typename?: 'Analytics';
  google_analytics?: Maybe<Scalars['String']>;
  umami_id?: Maybe<Scalars['String']>;
  simple_analytics?: Maybe<Scalars['String']>;
  plausible_data_domain?: Maybe<Scalars['String']>;
};

export type Author = {
  __typename?: 'Author';
  id: Scalars['Int'];
  email: Scalars['String'];
  username: Scalars['String'];
  name: Scalars['String'];
  social?: Maybe<Social>;
  bio?: Maybe<Scalars['String']>;
  occupation?: Maybe<Scalars['String']>;
  company_name?: Maybe<Scalars['String']>;
  role?: Maybe<Role>;
  permissions?: Maybe<Array<Permissions>>;
  avatar?: Maybe<Scalars['String']>;
  verified?: Maybe<Scalars['Boolean']>;
  accessToken?: Maybe<Scalars['String']>;
  analytics_id?: Maybe<Scalars['Int']>;
  analytics_uuid?: Maybe<Scalars['String']>;
};
// ...
export type CreatePostResponse = Post | PostError;
// ...
export type Image = {
  __typename?: 'Image';
  src?: Maybe<Scalars['String']>;
  width?: Maybe<Scalars['Int']>;
  height?: Maybe<Scalars['Int']>;
};
// ...
export type InputImage = {
  height?: Maybe<Scalars['Int']>;
  src: Scalars['String'];
  width?: Maybe<Scalars['Int']>;
};
// ...
export type InputUpdatePost = {
  id: Scalars['Int'];
  title?: Maybe<Scalars['String']>;
  html?: Maybe<Scalars['String']>;
  html_draft?: Maybe<Scalars['String']>;
  featured?: Maybe<Scalars['Boolean']>;
  excerpt?: Maybe<Scalars['String']>;
  cover_image?: Maybe<InputImage>;
  publishedAt?: Maybe<Scalars['Date']>;
  scheduledAt?: Maybe<Scalars['Date']>;
  updatedAt?: Maybe<Scalars['Date']>;
  type?: Maybe<PostTypes>;
  status?: Maybe<PostStatusOptions>;
  slug?: Maybe<Scalars['String']>;
  tags?: Maybe<Array<TagsInputType>>;
};

export type InputUpdateSubscriber = {
  secret_id?: Maybe<Scalars['String']>;
  verified?: Maybe<Scalars['Boolean']>;
};

export type LetterpadError = {
  message: Scalars['String'];
};
// ...
export type MutationUpdatePostArgs = {
  data?: Maybe<InputUpdatePost>;
};
// ...
export type Post = {
  __typename?: 'Post';
  id: Scalars['Int'];
  title: Scalars['String'];
  html?: Maybe<Scalars['String']>;
  html_draft?: Maybe<Scalars['String']>;
  author?: Maybe<Author>;
  excerpt?: Maybe<Scalars['String']>;
  cover_image: Image;
  type: PostTypes;
  status: PostStatusOptions;
  featured: Scalars['Boolean'];
  slug?: Maybe<Scalars['String']>;
  createdAt: Scalars['Date'];
  publishedAt?: Maybe<Scalars['Date']>;
  scheduledAt?: Maybe<Scalars['Date']>;
  updatedAt: Scalars['Date'];
  reading_time?: Maybe<Scalars['String']>;
  tags?: Maybe<Array<Tags>>;
};

export type PostCountsByStatus = {
  __typename?: 'PostCountsByStatus';
  published: Scalars['Int'];
  drafts: Scalars['Int'];
};

export type PostError = LetterpadError & {
  __typename?: 'PostError';
  message: Scalars['String'];
};

export type PostFilters = {
  id?: Maybe<Scalars['Int']>;
  slug?: Maybe<Scalars['String']>;
  featured?: Maybe<Scalars['Boolean']>;
  previewHash?: Maybe<Scalars['String']>;
  status?: Maybe<PostStatusOptions>;
  type?: Maybe<PostTypes>;
};

export type PostResponse = Post | PostError;

export enum PostStatusOptions {
  Published = 'published',
  Draft = 'draft',
  Trashed = 'trashed'
}

export enum PostTypes {
  Post = 'post',
  Page = 'page'
}
// ...
export type Tags = {
  __typename?: 'Tags';
  name: Scalars['String'];
  slug: Scalars['String'];
  posts?: Maybe<PostsResponse>;
};

export type TagsError = LetterpadError & {
  __typename?: 'TagsError';
  message: Scalars['String'];
};

export type TagsFilters = {
  active?: Maybe<Scalars['Boolean']>;
  name?: Maybe<Scalars['String']>;
  suggest?: Maybe<Scalars['Boolean']>;
};

export type TagsInputType = {
  name: Scalars['String'];
  slug: Scalars['String'];
};
// ...
export type ResolversTypes = {
// ...
  CreatePostResponse: ResolversTypes['Post'] | ResolversTypes['PostError'];
// ...
};